Updating lifespan tracking
